+++ junos/resource_system.go
@@ -515,6 +515,71 @@ func resourceSystem() *schema.Resource {
 								},
 							},
 						},
+						"web_management_http": {
+							Type:     schema.TypeList,
+							Optional: true,
+							MaxItems: 1,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"interface": {
+										Type:     schema.TypeList,
+										Optional: true,
+										Elem:     &schema.Schema{Type: schema.TypeString},
+									},
+									"port": {
+										Type:         schema.TypeInt,
+										Optional:     true,
+										ValidateFunc: validation.IntBetween(1, 65535),
+									},
+								},
+							},
+						},
+						"web_management_https": {
+							Type:     schema.TypeList,
+							Optional: true,
+							MaxItems: 1,
+							Elem: &schema.Resource{
+								Schema: map[string]*schema.Schema{
+									"interface": {
+										Type:     schema.TypeList,
+										Optional: true,
+										Elem:     &schema.Schema{Type: schema.TypeString},
+									},
+									"local_certificate": {
+										Type:     schema.TypeString,
+										Optional: true,
+										ExactlyOneOf: []string{
+											"services.0.web_management_https.0.local_certificate",
+											"services.0.web_management_https.0.pki_local_certificate",
+											"services.0.web_management_https.0.system_generated_certificate",
+										},
+									},
+									"pki_local_certificate": {
+										Type:     schema.TypeString,
+										Optional: true,
+										ExactlyOneOf: []string{
+											"services.0.web_management_https.0.local_certificate",
+											"services.0.web_management_https.0.pki_local_certificate",
+											"services.0.web_management_https.0.system_generated_certificate",
+										},
+									},
+									"port": {
+										Type:         schema.TypeInt,
+										Optional:     true,
+										ValidateFunc: validation.IntBetween(1, 65535),
+									},
+									"system_generated_certificate": {
+										Type:     schema.TypeBool,
+										Optional: true,
+										ExactlyOneOf: []string{
+											"services.0.web_management_https.0.local_certificate",
+											"services.0.web_management_https.0.pki_local_certificate",
+											"services.0.web_management_https.0.system_generated_certificate",
+										},
+									},
+								},
+							},
+						},
 					},
 				},
 			},
@@ -901,6 +966,40 @@ func setSystemServices(d *schema.ResourceData, m interface{}, jnprSess *NetconfO
 				return fmt.Errorf("services.0.ssh block is empty")
 			}
 		}
+		for _, http := range servicesM["web_management_http"].([]interface{}) {
+			configSet = append(configSet, setPrefix+"web-management http")
+			if http != nil {
+				httpOptions := http.(map[string]interface{})
+				for _, interf := range httpOptions["interface"].([]interface{}) {
+					configSet = append(configSet, setPrefix+"web-management http interface "+interf.(string))
+				}
+				if httpOptions["port"].(int) > 0 {
+					configSet = append(configSet, setPrefix+"web-management http port "+
+						strconv.Itoa(httpOptions["port"].(int)))
+				}
+			}
+		}
+		for _, https := range servicesM["web_management_https"].([]interface{}) {
+			httpsOptions := https.(map[string]interface{})
+			for _, interf := range httpsOptions["interface"].([]interface{}) {
+				configSet = append(configSet, setPrefix+"web-management https interface "+interf.(string))
+			}
+			if httpsOptions["local_certificate"].(string) != "" {
+				configSet = append(configSet,
+					setPrefix+"web-management https local-certificate \""+httpsOptions["local_certificate"].(string)+"\"")
+			}
+			if httpsOptions["pki_local_certificate"].(string) != "" {
+				configSet = append(configSet,
+					setPrefix+"web-management https pki-local-certificate \""+httpsOptions["pki_local_certificate"].(string)+"\"")
+			}
+			if httpsOptions["port"].(int) > 0 {
+				configSet = append(configSet, setPrefix+"web-management https port "+
+					strconv.Itoa(httpsOptions["port"].(int)))
+			}
+			if httpsOptions["system_generated_certificate"].(bool) {
+				configSet = append(configSet, setPrefix+"web-management https system-generated-certificate")
+			}
+		}
 	}
 
 	return sess.configSet(configSet, jnprSess)
@@ -1132,6 +1231,7 @@ func listLinesLogin() []string {
 func listLinesServices() []string {
 	ls := make([]string, 0)
 	ls = append(ls, listLinesServicesSSH()...)
+	ls = append(ls, listLinesServicesWebManagement()...)
 
 	return ls
 }
@@ -1158,6 +1258,13 @@ func listLinesServicesSSH() []string {
 		"services ssh root-login",
 		"services ssh no-tcp-forwarding",
 		"services ssh tcp-forwarding",
+	}
+}
+
+func listLinesServicesWebManagement() []string {
+	return []string{
+		"services web-management http",
+		"services web-management https",
 	}
 }
 
@@ -1286,11 +1393,18 @@ func readSystem(m interface{}, jnprSess *NetconfObject) (systemOptions, error) {
 			case checkStringHasPrefixInList(itemTrim, listLinesServices()):
 				if len(confRead.services) == 0 {
 					confRead.services = append(confRead.services, map[string]interface{}{
-						"ssh": make([]map[string]interface{}, 0),
+						"ssh":                  make([]map[string]interface{}, 0),
+						"web_management_http":  make([]map[string]interface{}, 0),
+						"web_management_https": make([]map[string]interface{}, 0),
 					})
 				}
 				if checkStringHasPrefixInList(itemTrim, listLinesServicesSSH()) {
 					if err := readSystemServicesSSH(&confRead, itemTrim); err != nil {
+						return confRead, err
+					}
+				}
+				if checkStringHasPrefixInList(itemTrim, listLinesServicesWebManagement()) {
+					if err := readSystemServicesWebManagement(&confRead, itemTrim); err != nil {
 						return confRead, err
 					}
 				}
@@ -1741,6 +1855,71 @@ func readSystemServicesSSH(confRead *systemOptions, itemTrim string) error {
 		confRead.services[0]["ssh"].([]map[string]interface{})[0]["no_tcp_forwarding"] = true
 	case itemTrim == "services ssh tcp-forwarding":
 		confRead.services[0]["ssh"].([]map[string]interface{})[0]["tcp_forwarding"] = true
+	}
+
+	return nil
+}
+
+func readSystemServicesWebManagement(confRead *systemOptions, itemTrim string) error {
+	switch {
+	case strings.HasPrefix(itemTrim, "services web-management https "):
+		if len(confRead.services[0]["web_management_https"].([]map[string]interface{})) == 0 {
+			confRead.services[0]["web_management_https"] = append(
+				confRead.services[0]["web_management_https"].([]map[string]interface{}),
+				map[string]interface{}{
+					"interface":                    make([]string, 0),
+					"port":                         0,
+					"local_certificate":            "",
+					"pki_local_certificate":        "",
+					"system_generated_certificate": false,
+				})
+		}
+		webMHTTPS := confRead.services[0]["web_management_https"].([]map[string]interface{})[0]
+		if strings.HasPrefix(itemTrim, "services web-management https interface ") {
+			webMHTTPS["interface"] = append(webMHTTPS["interface"].([]string),
+				strings.TrimPrefix(itemTrim, "services web-management https interface "))
+		}
+		if strings.HasPrefix(itemTrim, "services web-management https port ") {
+			var err error
+			webMHTTPS["port"], err =
+				strconv.Atoi(strings.TrimPrefix(itemTrim, "services web-management https port "))
+			if err != nil {
+				return fmt.Errorf("failed to convert value from '%s' to integer : %w", itemTrim, err)
+			}
+		}
+		if strings.HasPrefix(itemTrim, "services web-management https local-certificate ") {
+			webMHTTPS["local_certificate"] = strings.Trim(strings.TrimPrefix(itemTrim,
+				"services web-management https local-certificate "), "\"")
+		}
+		if strings.HasPrefix(itemTrim, "services web-management https pki-local-certificate ") {
+			webMHTTPS["pki_local_certificate"] = strings.Trim(strings.TrimPrefix(itemTrim,
+				"services web-management https pki-local-certificate "), "\"")
+		}
+		if itemTrim == "services web-management https system-generated-certificate" {
+			webMHTTPS["system_generated_certificate"] = true
+		}
+	case strings.HasPrefix(itemTrim, "services web-management http"):
+		if len(confRead.services[0]["web_management_http"].([]map[string]interface{})) == 0 {
+			confRead.services[0]["web_management_http"] = append(
+				confRead.services[0]["web_management_http"].([]map[string]interface{}),
+				map[string]interface{}{
+					"interface": make([]string, 0),
+					"port":      0,
+				})
+		}
+		webMHTTP := confRead.services[0]["web_management_http"].([]map[string]interface{})[0]
+		if strings.HasPrefix(itemTrim, "services web-management http interface ") {
+			webMHTTP["interface"] = append(webMHTTP["interface"].([]string),
+				strings.TrimPrefix(itemTrim, "services web-management http interface "))
+		}
+		if strings.HasPrefix(itemTrim, "services web-management http port ") {
+			var err error
+			webMHTTP["port"], err =
+				strconv.Atoi(strings.TrimPrefix(itemTrim, "services web-management http port "))
+			if err != nil {
+				return fmt.Errorf("failed to convert value from '%s' to integer : %w", itemTrim, err)
+			}
+		}
 	}
 
 	return nil